fix tests, halfway improve impl
